% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openentreprise.R
\name{extract_basicinfos}
\alias{extract_basicinfos}
\title{Extract basic infos}
\usage{
extract_basicinfos(x)
}
\arguments{
\item{x}{a get request to openentreprise}
}
\value{
a tibble
}
\description{
Extract basic infos
}
\examples{
get_openentreprise(.siren = "533735932") \%>\%
extract_basicinfos()

}
