% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openentreprise.R
\name{extract_mainoffice}
\alias{extract_mainoffice}
\title{Extract main office}
\usage{
extract_mainoffice(x)
}
\arguments{
\item{x}{a get request to openentreprise.fr}
}
\value{
a tibble
}
\description{
Extract main office
}
\examples{
get_openentreprise(.siren = "533735932") \%>\%
extract_mainoffice()

}
