% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openentreprise.R
\name{extract_url}
\alias{extract_url}
\title{Extract URL}
\usage{
extract_url(x)
}
\arguments{
\item{x}{a request to open API}
}
\value{
an URL as a string
}
\description{
extract the URL of the webpage
}
\examples{

extract_url(x = get_openentreprise(.siren = "533735932"))

\dontrun{
extract_url(x = get_openentreprise(.siren = "533735932")) \%>\%
browseURL()
}

}
