% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openentreprise.R
\name{get_financialdata}
\alias{get_financialdata}
\title{Get financial data}
\usage{
get_financialdata(.siren)
}
\arguments{
\item{.siren}{a valid siret number}
}
\value{
a tibble
}
\description{
Get financial data
}
\examples{

get_financialdata(.siren = "533735932")

}
