% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openentreprise.R
\name{get_mainoffice}
\alias{get_mainoffice}
\title{Get main office}
\usage{
get_mainoffice(.siren)
}
\arguments{
\item{.siren}{a siren number}
}
\description{
Get main office
}
\examples{
get_mainoffice(.siren = "533735932")

}
