% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openentreprise.R
\name{get_openentreprise}
\alias{get_openentreprise}
\title{Get Openentreprise}
\usage{
get_openentreprise(.siren)
}
\arguments{
\item{.siren}{a valid siren number}
}
\value{
a list
}
\description{
Get data from OpenEntreprise API : https://www.openentreprise.fr/swagger-ui.html
}
\examples{
get_openentreprise(.siren = "533735932")

}
