% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openentreprise.R
\name{get_url}
\alias{get_url}
\title{Get URL}
\usage{
get_url(.siren)
}
\arguments{
\item{.siren}{a siren number}
}
\value{
a string
}
\description{
A wrapper for extract_url()
}
\examples{
get_url(.siren = "533735932")

}
